pyproject.toml
[tool.poetry]
name = "ChatWebsite"
version = "0.1.0"
description = ""
authors = ["Dmitry Veko <blankvektrius@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
psycopg2-binary = "^2.9.3"
django-debug-toolbar = "^3.6.0"
Pillow = "^9.2.0"
django-guardian = "^2.4.0"
channels = "^3.0.5"
channels-redis = "^3.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
